REnamed node visitor
