+++ src/unused_finder/export_collector_tests.rs
@@ -11,9 +11,9 @@ mod test {
     use swc_ecma_parser::{Capturing, Parser};
 
     use crate::get_imports::create_lexer;
-    use crate::unused_finder::export_collector::{ExportedItem, ImportedItem};
+    use crate::unused_finder::node_visitor::{ExportedItem, ImportedItem};
 
-    use crate::unused_finder::export_collector::ExportsCollector;
+    use crate::unused_finder::node_visitor::ExportsCollector;
 
     fn create_test_parser<'a>(
         fm: &'a Arc<SourceFile>,
